+++ lib/commercetools-api/src/Models/Customer/CustomerCreatePasswordResetToken.php
@@ -15,6 +15,7 @@ interface CustomerCreatePasswordResetToken extends JsonObject
 {
     public const FIELD_EMAIL = 'email';
     public const FIELD_TTL_MINUTES = 'ttlMinutes';
+    public const FIELD_INVALIDATE_OLDER_TOKENS = 'invalidateOlderTokens';
 
     /**
      * <p>Email address of the Customer treated as <a href="/../api/customers-overview#email-case-insensitivity">case-insensitive</a>.</p>
@@ -33,6 +34,14 @@ interface CustomerCreatePasswordResetToken extends JsonObject
     public function getTtlMinutes();
 
     /**
+     * <p>If set to <code>true</code>, all password tokens issued previously for the Customer will be invalidated.</p>
+     *
+
+     * @return null|bool
+     */
+    public function getInvalidateOlderTokens();
+
+    /**
      * @param ?string $email
      */
     public function setEmail(?string $email): void;
@@ -41,4 +50,9 @@ interface CustomerCreatePasswordResetToken extends JsonObject
      * @param ?int $ttlMinutes
      */
     public function setTtlMinutes(?int $ttlMinutes): void;
+
+    /**
+     * @param ?bool $invalidateOlderTokens
+     */
+    public function setInvalidateOlderTokens(?bool $invalidateOlderTokens): void;
 }