+++ lib/commercetools-api/src/Models/Message/CustomerEmailTokenCreatedMessageBuilder.php
@@ -100,6 +100,18 @@ final class CustomerEmailTokenCreatedMessageBuilder implements Builder
     private $expiresAt;
 
     /**
+
+     * @var ?string
+     */
+    private $value;
+
+    /**
+
+     * @var ?bool
+     */
+    private $invalidateOlderTokens;
+
+    /**
      * <p>Unique identifier of the Message. Can be used to track which Messages have been processed.</p>
      *
 
@@ -230,6 +242,28 @@ final class CustomerEmailTokenCreatedMessageBuilder implements Builder
     public function getExpiresAt()
     {
         return $this->expiresAt;
+    }
+
+    /**
+     * <p>Value of the token, present only if the token's validity is 60 minutes or less.</p>
+     *
+
+     * @return null|string
+     */
+    public function getValue()
+    {
+        return $this->value;
+    }
+
+    /**
+     * <p>If <code>true</code>, all email tokens issued previously for the Customer are invalidated.</p>
+     *
+
+     * @return null|bool
+     */
+    public function getInvalidateOlderTokens()
+    {
+        return $this->invalidateOlderTokens;
     }
 
     /**
@@ -365,6 +399,28 @@ final class CustomerEmailTokenCreatedMessageBuilder implements Builder
     }
 
     /**
+     * @param ?string $value
+     * @return $this
+     */
+    public function withValue(?string $value)
+    {
+        $this->value = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param ?bool $invalidateOlderTokens
+     * @return $this
+     */
+    public function withInvalidateOlderTokens(?bool $invalidateOlderTokens)
+    {
+        $this->invalidateOlderTokens = $invalidateOlderTokens;
+
+        return $this;
+    }
+
+    /**
      * @deprecated use withLastModifiedBy() instead
      * @return $this
      */
@@ -422,7 +478,9 @@ final class CustomerEmailTokenCreatedMessageBuilder implements Builder
             $this->resourceVersion,
             $this->resourceUserProvidedIdentifiers instanceof UserProvidedIdentifiersBuilder ? $this->resourceUserProvidedIdentifiers->build() : $this->resourceUserProvidedIdentifiers,
             $this->customerId,
-            $this->expiresAt
+            $this->expiresAt,
+            $this->value,
+            $this->invalidateOlderTokens
         );
     }